fix(web): merge request-ID injection into proxy.ts for Next.js 16 compatibility
Next.js 16 replaced middleware.ts with proxy.ts as the sole intercept point
and raises a build error when both files coexist.  Merge the X-Request-Id
stamping logic from the deleted middleware.ts directly into proxy.ts, which
already owns auth-redirect logic.  Redirect responses also carry the header.
